requirements.txt
fastmcp
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
openpyxl
pandas
aiofiles
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
playwright
requests
streamlit
langchain
langchain-openai
langchain-google-genai
langchain-community
pydantic
openai
schedule
